requirements.txt
Flask
mysql-connector-python
bcrypt
gunicorn
python-dotenv